move config.go into its own package
