+++ traffic_ops/traffic_ops_golang/config/config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
