+++ traffic_ops/traffic_ops_golang/routing.go
@@ -32,6 +32,7 @@ import (
 	"fmt"
 
 	"github.com/apache/incubator-trafficcontrol/traffic_ops/traffic_ops_golang/api"
+	"github.com/apache/incubator-trafficcontrol/traffic_ops/traffic_ops_golang/config"
 	"github.com/jmoiron/sqlx"
 )
 
@@ -59,7 +60,7 @@ func getDefaultMiddleware() []Middleware {
 
 // ServerData ...
 type ServerData struct {
-	Config
+	config.Config
 	DB *sqlx.DB
 }
